Вроде работает кроме клика по удалению второй раз

+++ app/src/main/java/com/sloot/shoppingdal/ItemShop.java
@@ -1,10 +1,12 @@
 package com.sloot.shoppingdal;
 
+
 public class ItemShop {
-    public int id;
+    public long id;
     public String name;
     public boolean buy;
-    public ItemShop(int id, String name, boolean buy) {
+
+    public ItemShop(long id, String name, boolean buy) {
         this.id = id;
         this.name = name;
         this.buy = buy;
